Go to day in calendar, if previous/next not available in log
If there's no previous/next date in the daily log archive (which in most
cases means there's likely no log archive for the day at all), then show
a button that goes to the previous/next day in the calendar instead.
Except if the next day is in the future.

refs #4